requirements/base.txt
# Core dependencies for all environments
Django~=4.2.8
asgiref~=3.6.0
backoff~=2.2.1
boto3~=1.37.4
django-cachalot~=2.6.1
django-environ~=0.11.2
django-filter~=23.5
django-storages~=1.14.5
djangorestframework~=3.14.0
gunicorn~=21.2.0
numpy~=1.26.2
pandas~=2.1.4
psycopg[binary]>=3.2.12
python-jose~=3.4.0
requests~=2.31.0
sentry-sdk~=2.0.0
tqdm~=4.66.1
